Moved parse-user-agent into the processors dir
