+++ src/services/proxy/processors/parse-user-agent.ts
@@ -1,5 +1,5 @@
 import uap from 'ua-parser-js';
-import {FastifyRequest} from '../../types';
+import {FastifyRequest} from '../../../types';
 
 function isBot(userAgentString: string): boolean {
     const botPattern = /wget|ahrefsbot|curl|bot|crawler|spider|urllib|bitdiscovery|\+https:\/\/|googlebot/i;
