+++ test/services/proxy/processors/parse-user-agent.ts
@@ -1,6 +1,6 @@
 import {describe, it, expect} from 'vitest';
-import * as processors from '../../../src/services/proxy/processors';
-import {FastifyRequest, HttpProxyRequest} from '../../../src/types';
+import * as processors from '../../../../src/services/proxy/processors/parse-user-agent';
+import {FastifyRequest, HttpProxyRequest} from '../../../../src/types';
 
 const {parseUserAgent} = processors;
